refactor(jobs): update import paths for job schemas
