+++ lib/types/jobs/schemas/base-job.schema.ts
